Create url_agent_bypass.py
Aug 30, 2023

This is a script that will bypass blocks that come in the form of something like: 403 Forbidden. Essentially, this code refers to a server refusing a connection despite understanding the nature and intent of the request being made. 

Servers rely on headers presented in the way of user agents e.g., Chrome, Firefox, and the devices that pertain to them: Chromebook, iPad, Linux, Windows etc. My script simply tells the server that the client sending the request is an iPad and the server being none the wiser accepts the request and transmits the data accordingly. 

This was written as a quick hack. The libraries used here are standard. 

NOTE: The URL included here is for a Kali image and that's a good image to download if you are testing this. Trust that the good folks over at Offensive Security not only are fine with you testing Python scripts against their servers, they likely encourage it. What better things are there to do for aspiring and seasoned pentesters alike.